+++ src/components/index.ts
@@ -1,5 +1,5 @@
 export { default as Typography } from "./Typography/Typography";
 export * from "./Typography/Typography";
 
-export { default as WeatherSection } from "./WeatherSection/WeatherSection";
-export { default as NewsSection } from "./NewsSection/NewsSection";
+export { default as WeatherBoxes } from "./WeatherBoxes/WeatherBoxes";
+export { default as DailyReport } from "./DailyReport/DailyReport";
