+++ src/views/index.ts
@@ -1,0 +1,2 @@
+export { default as WeatherSection } from "./WeatherSection/WeatherSection";
+export { default as NewsSection } from "./NewsSection/NewsSection";
